fix lexing of comments
previously, the trailing '\n' of a comment would only advance the input
by 1 character, but if the '\n' is at the end of a comment, we have to
skip the whole comment. this makes sure that happens. somewhat nasty,
but at least the issue was found and there is a test now.